pyproject.toml
[project]
name = "tech-radar-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.16",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-litellm>=0.2.2",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.30",
    "langchain-text-splitters>=0.3.9",
    "litellm==1.74.8",
    "pypdf==5.8.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
seed-db = "src.scripts.db_management:migrate_and_seed"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.setuptools]
packages = [
    "src",
]
